+++ examples/example.ts
@@ -79,7 +79,7 @@ const input = { // input values
 	remoteAddrIpv6: "::1",
 	limit: "0",
 };
-const expected = { // should be converted to this
+const expected = { // should be transformed to this
 	id: 1,
 	name: "Pablo Diego José",
 	birthday: new Date("2000-01-02T03:04:05.678Z"),