+++ src/schemaClasses/ArraySchema.ts
@@ -1,7 +1,8 @@
-import * as converter from "../appliers/converter";
 import * as ifEmptyString from "../appliers/ifEmptyString";
 import * as ifNull from "../appliers/ifNull";
 import * as ifUndefined from "../appliers/ifUndefined";
+import * as transform from "../appliers/transform";
+
 import * as each from "../appliers/array/each";
 import * as maxLength from "../appliers/array/maxLength";
 import * as minLength from "../appliers/array/minLength";
@@ -10,7 +11,7 @@ import * as type from "../appliers/array/type";
 import {BaseSchema} from "./BaseSchema";
 
 export type RulesForArray<T> =
-	converter.Rules<T[]> &
+	transform.Rules<T[]> &
 	ifUndefined.Rules<T[]> &
 	ifEmptyString.Rules<T[]> &
 	ifNull.Rules<T[]> &
@@ -31,7 +32,7 @@ export class ArraySchema<T, Tx = never> extends BaseSchema<T[] | Tx>
 			each.applyTo,
 			minLength.applyTo,
 			maxLength.applyTo,
-			converter.applyTo,
+			transform.applyTo,
 		]);
 	}
 }