+++ src/schemas/object.ts
@@ -1,6 +1,7 @@
+import {ObjectSchema, RulesForObject} from "../schemaClasses/ObjectSchema";
+
 import {NullableRules, UndefinableRules} from "../libs/publicTypes";
 import {SchemaObject} from "../libs/types";
-import {ObjectSchema, RulesForObject} from "../schemaClasses/ObjectSchema";
 
 /** schema for object or null or undefined */
 export function object<S extends SchemaObject>(options: RulesForObject<S> & NullableRules & UndefinableRules): ObjectSchema<S, null | undefined>;